GameData: add a new() constructor
Although we construct only a single instance of the structure, in a
single place of the code, it is rather idiomatic for a structure to have
a new() constructor.  Furthermore, the new constructor unclutters the
main() by having to pass only the necessary initial data to it and leave
the rest of the details to the constructor.