format_masked_string: convert function to method
First of all, format_masked_string() operates solely on a `GameData'
state.  By converting it into a method we loose some generality, but the
actual use of it is greatly simplified.

If we still wished to have the "general" version of the function,
we could keep it as a function and create a new method, a wrapper to
the function.  Since the program is rather small and we don't need this
option, we can avoid it for now.