update_screen: turn String parameter to &str
In general, it is a good practice functions to accept &str and
optionally return String.  The reason behind this practice is that a
String is implicitly converted (coerced) into a &str without any
penalty, whereas the opposite needs some heap allocation, sometimes
unnecessary in the context of a function.

In this case it is not such a big difference, as we already use Strings.
But, it is a good practice, nevertheless!