main: inverse order of conditional
It is a good style to keep the short part of a conditional close to the
`if' statement, so that both parts are closer at the end, even though we
introduce an inverse logic expression.  We could fix the inverse logic
by renaming the validate_user_guess(), maybe to something like
invalid_guess().

Furthermore, although in contrast to the logic of the program, it is
better to first handle the exceptional cases and then the "happy path".