main: place game conditions check at start
Although this logic seems "unnatural" to the game's course (first play,
then check the result vs. first check the result, then play), but it
will help us to remove one major code duplication: creating the masked
output and printing the status screen.

At the first iteration, both checks are ignored, as their conditions are
guaranteed not to hold.  After the first round, the conditions will be
examined at the second loop, instead of the end of the first loop.

Similarly, the new code is equivalent to the old one for the other
rounds:  we will run an extra loop cycle for the end-of-game conditions,
but the same number of rounds as the previous version of the code.
Only, that we will have less functions to call manually.